+++ example/tutorial/step3/main.js
@@ -1,7 +1,7 @@
 import { createBoard, playMove } from "./connect4.js";
 
 function getWebSocketServer() {
-  if (window.location.host === "aaugustin.github.io") {
+  if (window.location.host === "python-websockets.github.io") {
     return "wss://websockets-tutorial.herokuapp.com/";
   } else if (window.location.host === "localhost:8000") {
     return "ws://localhost:8001/";